+++ R/mapbayest_check_and_preprocess.R
@@ -60,21 +60,26 @@ check_mapbayr_model <- function(x){
 }
 
 check_mapbayr_data <- function(data){
+  # Is there any data?
   if(is.null(data)) stop("No data provided", call. = F)
 
+  # Are all column numerics
+  non_num <- names(data)[!purrr::map_lgl(data, is.numeric)]
+  if(length(non_num)) stop(paste("Non-numeric column found:", paste(non_num, collapse = " ")), call. = F)
+
+  # Are required items present?
   data <- data %>%
     rename_with(tolower, any_of(c("TIME", "AMT", "MDV", "CMT", "EVID", "II", "ADDL", "SS", "RATE")))
 
-  if(is.null(data[["ID"]]))   stop('ID column is missing', call. = F)
-  if(is.null(data[["time"]])) stop('time column is missing', call. = F)
-  if(is.null(data[["evid"]])) stop('evid column is missing', call. = F)
-  if(is.null(data[["cmt"]]))  stop('cmt column is missing', call. = F)
-  if(is.null(data[["amt"]]))  stop('amt column is missing', call. = F)
-  if(is.null(data[["DV"]]))   stop('DV column is missing', call. = F)
+  required_nmtran_item <- c("ID", "time", "evid", "cmt", "amt", "DV")
+  miss_item <- required_nmtran_item[!(required_nmtran_item %in% names(data))]
+  if(length(miss_item)) stop(paste("Missing column:", paste(miss_item, collapse = " ")), call. = F)
   if(is.null(data[["mdv"]])){
     data[["mdv"]] <- ifelse(data[["evid"]] %in% c(1,2,4), 1, 0)
   }
   if(is.null(data[["mdv"]]))  stop('mdv column is missing', call. = F) #Cannot happen obviously... but who knows
+
+  # Are MDV/EVID requirements respected?
 
   if(nrow(filter(data, .data$mdv == 0 & .data$evid == 2)) > 0) stop("Lines with evid = 2 & mdv = 0 are not allowed", call. = F)
   if(nrow(filter(data, .data$mdv == 0 & .data$evid != 0)) > 0) stop("Lines with mdv = 0 must have evid = 0.", call. = F)